Update to WirelessSensor complete
