+++ LibreNMS/OS/Ray.php
@@ -26,7 +26,7 @@
 namespace LibreNMS\OS;
 
 use LibreNMS\Device\Processor;
-use LibreNMS\Device\WirelessSensor;
+use App\Models\WirelessSensor;
 use LibreNMS\Interfaces\Discovery\ProcessorDiscovery;
 use LibreNMS\Interfaces\Discovery\Sensors\WirelessFrequencyDiscovery;
 use LibreNMS\Interfaces\Discovery\Sensors\WirelessPowerDiscovery;
